# Conflicts:
#	fangyuanyouyue-microservice/fangyuanyouyue-microservice-forum/src/main/java/com/fangyuanyouyue/forum/controller/ForumLikesController.java

+++ fangyuanyouyue-microservice/fangyuanyouyue-microservice-forum/src/main/java/com/fangyuanyouyue/forum/service/impl/ForumInfoServiceImp.java
@@ -39,10 +39,10 @@ public class ForumInfoServiceImp implements ForumInfoService {
 		if(forumInfo!=null) {
 			ForumInfoDto dto = new ForumInfoDto(forumInfo);
 			//计算点赞数
-			long likesCount = forumLikesService.countLikes(id);
+			Integer	 likesCount = forumLikesService.countLikes(id);
 			dto.setLikesCount(likesCount);
 			//计算评论数
-			long commentCount = forumCommentService.countComment(id);
+			Integer commentCount = forumCommentService.countComment(id);
 			dto.setCommentCount(commentCount);
 
 			//TODO 如果不是视频，需要找到图片
@@ -81,6 +81,7 @@ public class ForumInfoServiceImp implements ForumInfoService {
 				}else{
 					forumInfo.setColumnId(columnId);
 				}
+				//TODO 社交消息：您的专栏【专栏标题】有新的帖子，点击此处前往查看吧
 			}
 		}else if(type == 2){//视频
 			if(StringUtils.isEmpty(videoUrl) || StringUtils.isEmpty(videoImg) || videoLength == null){
@@ -94,7 +95,8 @@ public class ForumInfoServiceImp implements ForumInfoService {
 		}
 		forumInfoMapper.insert(forumInfo);
 		if(userIds != null && userIds.length > 0){
-			//TODO 发送通知邀请用户
+			//TODO 发送通知邀请用户 邀请我：用户“用户昵称”上传帖子【帖子名称】时邀请了您！点击此处前往查看吧
+			//邀请我：用户“用户昵称”上传视频【视频名称】时邀请了您！点击此处前往查看吧
 			for(Integer toUserId:userIds){
 
 			}
